Refactored Command processor and Command Processor service as example
