--- src/Paramore.Brighter.Extensions.DependencyInjection/IAmAScopedCommandProcessor.cs
@@ -1,5 +1,0 @@
-﻿namespace Paramore.Brighter.Extensions.DependencyInjection
-{
-    public interface IAmAScopedCommandProcessor : IAmACommandProcessor
-    { }
-}
